+++ kolibri/plugins/learn/assets/src/vue/content-grid-item/images/default_thumbnail.png
