+++ kolibri/plugins/learn/assets/src/vue/content-grid-item/index.vue
@@ -1,7 +1,7 @@
 <template>
 
   <div>
-    <card v-link="link" :title="title">
+    <grid-item v-link="link" :title="title">
       <div class='thumbnail' :class="{ 'thumbnail-center' : !thumb }" :style='{ "background-image": thumb }'>
         <content-icon
           :class='thumb ? "content-icon" : "content-icon-center" '
@@ -10,7 +10,7 @@
           :progress="progress">
         </content-icon>
       </div>
-    </card>
+    </grid-item>
   </div>
 
 </template>
@@ -26,7 +26,7 @@
   module.exports = {
     components: {
       'content-icon': require('../content-icon'),
-      'card': require('../card'),
+      'grid-item': require('../card-grid/grid-item'),
     },
     props: {
       id: {
