+++ kolibri/plugins/learn/assets/src/vue/explore-page/index.vue
@@ -21,18 +21,18 @@
 
     <span class="visuallyhidden" v-if="subtopics.length">You can navigate groups of content through headings.</span>
 
-    <list-grid v-if="subtopics.length">
-      <topic-card
+    <card-list v-if="subtopics.length">
+      <topic-list-item
         v-for="topic in subtopics"
         :id="topic.id"
         :title="topic.title"
         :ntotal="topic.n_total"
         :ncomplete="topic.n_complete">
-      </topic-card>
-    </list-grid>
+      </topic-list-item>
+    </card-list>
 
     <card-grid v-if="contents.length">
-      <content-card
+      <content-grid-item
         v-for="content in contents"
         class="card"
         :title="content.title"
@@ -40,7 +40,7 @@
         :kind="content.kind"
         :progress="content.progress"
         :id="content.id">
-      </content-card>
+      </content-grid-item>
     </card-grid>
 
   </div>
@@ -54,10 +54,10 @@
     components: {
       'breadcrumbs': require('../breadcrumbs'),
       'page-header': require('../page-header'),
-      'topic-card': require('../topic-card'),
-      'content-card': require('../content-card'),
+      'topic-list-item': require('../topic-list-item'),
+      'content-grid-item': require('../content-grid-item'),
       'card-grid': require('../card-grid'),
-      'list-grid': require('../list-grid'),
+      'card-list': require('../card-list'),
     },
     computed: {
       title() {
